# Conflicts:
#	BlazorWASM-AppLogic/Application/GrpcClients/PostGrpcClient.cs
#	BlazorWASM-AppLogic/Application/Logic/PostLogic.cs
#	BlazorWASM-AppLogic/Application/LogicInterfaces/IPostLogic.cs
#	BlazorWASM-AppLogic/Application/obj/Debug/net6.0/Protos/Post.cs
#	BlazorWASM-AppLogic/Application/obj/Debug/net6.0/Protos/PostGrpc.cs
#	BlazorWASM-AppLogic/BlazorWASM/Pages/PostDetails.razor
#	BlazorWASM-AppLogic/BlazorWASM/Pages/ViewPosts.razor
#	BlazorWASM-AppLogic/HttpClients/ClientInterfaces/IPostService.cs
#	BlazorWASM-AppLogic/HttpClients/Implementations/PostHttpClient.cs
#	BlazorWASM-AppLogic/WebAPI/Services/AuthService.cs
#	DataPersistence/src/main/java/sep3/project/daos/impl/PostDatabase.java
#	DataPersistence/src/main/java/sep3/project/daos/impl/UserDatabase.java
#	DataPersistence/src/main/java/sep3/project/protobuf/Post.java
#	DataPersistence/src/main/proto/post.proto

+++ BlazorWASM-AppLogic/Domain/DTOs/LikePostDto.cs
@@ -2,10 +2,10 @@
 
 public class LikePostDto
 {
-    public int postId { get; }
+    public long postId { get; }
     public long userId { get; }
 
-    public LikePostDto(int postId, long userId)
+    public LikePostDto(long postId, long userId)
     {
         this.postId = postId;
         this.userId = userId;
