+++ BlazorWASM-AppLogic/Domain/Models/Like.cs
@@ -2,10 +2,10 @@
 
 public class Like
 {
-    public int likedBy { get; set; }
+    public long likedBy { get; set; }
     public long likedPost { get; set; }
 
-    public Like(int likedBy, long likedPost)
+    public Like(long likedBy, long likedPost)
     {
         this.likedBy = likedBy;
         this.likedPost = likedPost;
